requirements.txt
requests~=2.25.0
pika~=1.2.0
coloredlogs~=15.0.1
python-dotenv~=0.19.1